Updated the id type from Long to Integer according to Documentation.

+++ src/main/java/com/cooksys/June2020/dtos/TweetResponseDto.java
@@ -1,26 +1,26 @@
 package com.cooksys.June2020.dtos;
 
+import java.sql.Timestamp;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.sql.Timestamp;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
 public class TweetResponseDto {
-    private Long id;
+	private Integer id;
 
-    private UserResponseDto author;
+	private UserResponseDto author;
 
-    private Timestamp timestamp;
+	private Timestamp timestamp;
 
-    private String content;
+	private String content;
 
-    private TweetResponseDto inReplyTo;
+	private TweetResponseDto inReplyTo;
 
-    private TweetResponseDto repostOf;
+	private TweetResponseDto repostOf;
 }
