+++ src/main/java/com/cooksys/June2020/entities/HashTag.java
@@ -1,15 +1,23 @@
 package com.cooksys.June2020.entities;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import java.sql.Timestamp;
+import java.util.List;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToMany;
+import javax.persistence.Table;
 
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import javax.persistence.*;
-import java.sql.Timestamp;
-import java.util.List;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Entity
 @Table(name = "hashtags")
@@ -17,23 +25,25 @@ import java.util.List;
 @Getter
 @Setter
 public class HashTag {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.SEQUENCE)
+	private Integer id;
 
-    // HashTag parsing from a tweet's contents should lowercase the tags when tagging
-    // Since all hashtags will be converted to lowercase when contents are parsed, they'll be unique!
-    @Column(unique = true, nullable = false)
-    private String label;
+	// HashTag parsing from a tweet's contents should lowercase the tags when
+	// tagging
+	// Since all hashtags will be converted to lowercase when contents are parsed,
+	// they'll be unique!
+	@Column(unique = true, nullable = false)
+	private String label;
 
-    @CreationTimestamp
-    @Column(nullable = false)
-    private Timestamp firstUsed;
+	@CreationTimestamp
+	@Column(nullable = false)
+	private Timestamp firstUsed;
 
-    @UpdateTimestamp
-    @Column(nullable = false)
-    private Timestamp lastUsed;
+	@UpdateTimestamp
+	@Column(nullable = false)
+	private Timestamp lastUsed;
 
-    @ManyToMany(cascade = CascadeType.ALL, mappedBy = "hashtags")
-    private List<Tweet> tweets;
+	@ManyToMany(cascade = CascadeType.ALL, mappedBy = "hashtags")
+	private List<Tweet> tweets;
 }
